refactor: change nam for filter slice and path & connected category to main screen
