+++ src/components/filters/subComponents/filtersButtonComponent.js
@@ -1,14 +1,19 @@
 import { View, TouchableOpacity, StyleSheet } from "react-native";
 import Entypo from "@expo/vector-icons/Entypo";
 import { useDispatch } from "react-redux";
-import { showModal } from "../../../store/slices/filterSlice";
+import { showModal } from "../../../store/slices/filter.slice";
 
 export const FiltersButtonComponent = () => {
   const dispatch = useDispatch();
 
   return (
     <View style={styles.container}>
-      <TouchableOpacity onPress={() => dispatch(showModal())}>
+      <TouchableOpacity
+        onPress={() => {
+          console.log("Button clicked");
+          dispatch(showModal());
+        }}
+      >
         <Entypo name="sound-mix" size={30} color="#F7CCC3" />
       </TouchableOpacity>
     </View>
