+++ src/components/menu/NameMenuComponent.js
@@ -1,7 +1,7 @@
 import { StyleSheet, View, Text } from "react-native";
 import { PopMenuComponent } from "./PopMenuComponent";
 import { SearchBarComponent } from "../searchBar/SearchBarComponent.js";
-import { FiltersButtonComponent } from "../filters/subComponents/FiltersButtonComponent.js";
+import { FiltersButtonComponent } from "../filters/subComponents/FiltersButtonComponent";
 
 export const NameMenuComponent = ({ userName }) => {
   return (
