+++ src/components/ui/CategoriesList.js
@@ -1,10 +1,11 @@
-import React from "react";
 import { FlatList, View, StyleSheet } from "react-native";
 import CategoryComponent from "./CategoryComponent";
 import categories from "../../data/categories";
+import { HeadingComponent } from "../heading/HeadingComponent";
 
 const CategoriesList = () => (
   <View style={styles.container}>
+    <HeadingComponent children={"Category"} level={3} color="#000000" />
     <FlatList
       data={categories}
       horizontal
@@ -19,9 +20,10 @@ const CategoriesList = () => (
 const styles = StyleSheet.create({
   container: {
     paddingVertical: 16,
+    paddingHorizontal: 16,
+    gap: 20,
   },
   listContent: {
-    paddingHorizontal: 16,
     gap: 12,
   },
 });
